requirements.txt
streamlit
streamlit-extras
yfinance
pandas
numpy
altair
plotly
google-generativeai
matplotlib
requests
beautifulsoup4
python-dotenv
scikit-learn
seaborn
pillow
finnhub-python
ta
streamlit-option-menu